src:ticket:789
git-svn-id: https://www.myroms.org/svn/src/trunk@927 f091316a-d328-0410-a40a-876eff57d070

+++ ROMS/Drivers/ad_ocean.h
@@ -215,7 +215,7 @@
 !
 !  Imported variable declarations.
 !
-      real(r8), intent(in) :: RunInterval            ! seconds
+      real(dp), intent(in) :: RunInterval            ! seconds
 !
 !  Local variable declarations.
 !
